feature: you can now relay to a list of topics
You can now set up an array of nsq relays, with the restriction
that the event name must be the same as the nsq topic being posted to.
The following code snippet sets up relays to three different nsq
topics.

```js
const relayer = createRelayer({
   nsq: 'http://localhost:4151',
   relays: [ 'metric', 'email', 'cute-kitten' ],
});
```

Added tests for the new option and for the listener cleanup.
Documented the new option in the readme.

Test coverage is now missing one line (the success debug log).